gcore-shaders: move `registry::types`
