+++ node-graph/gcore-shaders/src/lib.rs
@@ -3,6 +3,7 @@ pub use glam;
 pub mod blending;
 pub mod choice_type;
 pub mod color;
+pub mod registry;
 
 pub trait AsU32 {
 	fn as_u32(&self) -> u32;
